requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
python-docx>=0.8.11
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
